Added index.d.ts in files option of package.json
In order for the types file to be included in the package as you can see
here: https://docs.npmjs.com/files/package.json#files. Without adding
index.d.ts, it will not be included in the package. So, I have added it
so it becomes available for people who use typescript.